assessment/requirements.txt
asgiref == 3.7.2
Django == 4.2.8
django-extensions == 3.2.3
django-shell-plus == 1.1.7
djangorestframework == 3.14.0
djangorestframework-simplejwt == 5.3.1
psycopg2-binary == 2.9.9
PyJWT == 2.8.0
pytz == 2023.3.post1
sqlparse == 0.4.4
typing_extensions == 4.9.0
